Репозитории, ошибки, интерфейсы, сущности, конфиги
