+++ telegram-bot/src/main/java/com/example/catphototg/TelegramBotApplication.java
@@ -1,0 +1,11 @@
+package com.example.catphototg;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TelegramBotApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(TelegramBotApplication.class, args);
+    }
+}
