--- telegram-bot/src/main/java/com/example/catphototg/bot/TelegramBotApplication.java
@@ -1,11 +1,0 @@
-package com.example.catphototg.bot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TelegramBotApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(TelegramBotApplication.class, args);
-    }
-}
