--- telegram-bot/src/main/java/com/example/catphototg/bot/config/AsyncConfig.java
@@ -1,21 +1,0 @@
-package com.example.catphototg.bot.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-import java.util.concurrent.Executor;
-
-@Configuration
-public class AsyncConfig {
-
-    @Bean(name = "asyncTaskExecutor")
-    public Executor asyncTaskExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(10);
-        executor.setMaxPoolSize(50);
-        executor.setQueueCapacity(100);
-        executor.setThreadNamePrefix("Async-Executor-");
-        executor.initialize();
-        return executor;
-    }
-}
