--- telegram-bot/src/main/java/com/example/catphototg/bot/config/BotConfig.java
@@ -1,27 +1,0 @@
-package com.example.catphototg.bot.config;
-
-import com.example.catphototg.bot.bot.CatBot;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-
-@Configuration
-public class BotConfig {
-
-    private final CatBot catBot;
-
-    @Autowired
-    public BotConfig(CatBot catBot) {
-        this.catBot = catBot;
-    }
-
-    @Bean
-    public TelegramBotsApi telegramBotsApi() throws TelegramApiException {
-        TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
-        botsApi.registerBot(catBot);
-        return botsApi;
-    }
-}
