--- telegram-bot/src/main/java/com/example/catphototg/bot/config/BotProperties.java
@@ -1,13 +1,0 @@
-package com.example.catphototg.bot.config;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@ConfigurationProperties(prefix = "bot")
-@Data
-public class BotProperties {
-    private String token;
-    private String name;
-}
