--- telegram-bot/src/main/java/com/example/catphototg/bot/entity/User.java
@@ -1,37 +1,0 @@
-package com.example.catphototg.bot.entity;
-
-import com.example.catphototg.catservice.entity.Cat;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "users")
-@Getter
-@Setter
-@NoArgsConstructor
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "telegram_id", unique = true, nullable = false)
-    private Long telegramId;
-
-    @Column(name = "username")
-    private String username;
-
-    @Column(name = "display_name")
-    private String displayName;
-
-    @OneToMany(mappedBy = "author", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<Cat> cats = new ArrayList<>();
-
-    public String displayName() {
-        return displayName;
-    }
-}
