--- telegram-bot/src/main/java/com/example/catphototg/bot/entity/ui/Button.java
@@ -1,4 +1,0 @@
-package com.example.catphototg.bot.entity.ui;
-
-public record Button (String text, String action){
-}
