--- telegram-bot/src/main/java/com/example/catphototg/bot/entity/ui/Keyboard.java
@@ -1,6 +1,0 @@
-package com.example.catphototg.bot.entity.ui;
-
-import java.util.List;
-
-public record Keyboard(List<KeyboardRow> rows) {
-}
