--- telegram-bot/src/main/java/com/example/catphototg/bot/entity/ui/MessageData.java
@@ -1,4 +1,0 @@
-package com.example.catphototg.bot.entity.ui;
-
-public record MessageData(String text, Keyboard keyboard) {
-}
