--- telegram-bot/src/main/java/com/example/catphototg/bot/exceptions/BotOperationException.java
@@ -1,7 +1,0 @@
-package com.example.catphototg.bot.exceptions;
-
-public class BotOperationException extends RuntimeException {
-    public BotOperationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
