--- telegram-bot/src/main/java/com/example/catphototg/bot/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.catphototg.bot.repository;
-
-import com.example.catphototg.bot.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByTelegramId(Long telegramId);
-
-}
