+++ telegram-bot/src/main/java/com/example/catphototg/config/BotConfig.java
@@ -1,0 +1,27 @@
+package com.example.catphototg.config;
+
+import com.example.catphototg.bot.CatBot;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.telegram.telegrambots.meta.TelegramBotsApi;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
+
+@Configuration
+public class BotConfig {
+
+    private final CatBot catBot;
+
+    @Autowired
+    public BotConfig(CatBot catBot) {
+        this.catBot = catBot;
+    }
+
+    @Bean
+    public TelegramBotsApi telegramBotsApi() throws TelegramApiException {
+        TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
+        botsApi.registerBot(catBot);
+        return botsApi;
+    }
+}
