+++ telegram-bot/src/main/java/com/example/catphototg/entity/User.java
@@ -1,0 +1,28 @@
+package com.example.catphototg.entity;
+
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "users")
+@Getter
+@Setter
+@NoArgsConstructor
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "telegram_id", unique = true, nullable = false)
+    private Long telegramId;
+
+    @Column(name = "username")
+    private String username;
+
+    @Column(name = "display_name")
+    private String displayName;
+
+}
