+++ telegram-bot/src/main/java/com/example/catphototg/entity/enums/UserState.java
@@ -1,0 +1,13 @@
+package com.example.catphototg.entity.enums;
+
+public enum UserState {
+    MAIN_MENU,
+    ADDING_CAT_NAME,
+    ADDING_CAT_PHOTO,
+    ADDING_CAT_CONFIRMATION,
+    BROWSING_MY_CATS,
+    REGISTERING_NAME,
+    VIEWING_CAT_DETAILS,
+    VIEWING_RANDOM_CAT,
+    CHANGING_NAME
+}
