+++ telegram-bot/src/main/java/com/example/catphototg/entity/ui/Button.java
@@ -1,0 +1,4 @@
+package com.example.catphototg.entity.ui;
+
+public record Button (String text, String action){
+}
