+++ telegram-bot/src/main/java/com/example/catphototg/entity/ui/KeyboardRow.java
@@ -1,0 +1,6 @@
+package com.example.catphototg.entity.ui;
+
+import java.util.List;
+
+public record KeyboardRow(List<Button> buttons) {
+}
