+++ telegram-bot/src/main/java/com/example/catphototg/entity/ui/MessageData.java
@@ -1,0 +1,4 @@
+package com.example.catphototg.entity.ui;
+
+public record MessageData(String text, Keyboard keyboard) {
+}
