+++ telegram-bot/src/main/java/com/example/catphototg/exceptions/BotOperationException.java
@@ -1,0 +1,7 @@
+package com.example.catphototg.exceptions;
+
+public class BotOperationException extends RuntimeException {
+    public BotOperationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
