+++ telegram-bot/src/main/java/com/example/catphototg/exceptions/NoSuchUserException.java
@@ -1,0 +1,8 @@
+package com.example.catphototg.exceptions;
+
+
+public class NoSuchUserException extends RuntimeException{
+    public NoSuchUserException(String message) {
+        super(message);
+    }
+}
