+++ telegram-bot/src/main/java/com/example/catphototg/handlers/interfaces/UpdateHandler.java
@@ -1,0 +1,11 @@
+package com.example.catphototg.handlers.interfaces;
+
+
+import com.example.catphototg.dto.TelegramMessage;
+import com.example.catphototg.entity.User;
+import com.example.catphototg.entity.UserSession;
+
+public interface UpdateHandler {
+    boolean canHandle(User user, UserSession session, TelegramMessage message);
+    void handle(User user, UserSession session, TelegramMessage message);
+}
