+++ telegram-bot/src/main/java/com/example/catphototg/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.catphototg.repository;
+
+import com.example.catphototg.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByTelegramId(Long telegramId);
+
+}
